Resolve connection issue in mongodb in controllers and routes
