--- NODEHTTP/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "api-masterclass",
-  "version": "1.0.0",
-  "description": "\"API master class with Node js\"",
-  "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon server.js"
-  },
-  "author": "Indrajith Bodhinyaka",
-  "license": "ISC",
-  "devDependencies": {
-    "nodemon": "^3.1.4"
-  }
-}
