--- devCamper-API/config/config.env
@@ -1,3 +1,0 @@
-NODE_ENV=development
-PORT=5000
-MONGO_URI=mongodb+srv://indrajithbodinayaka352:GRLbtAXHk8pIDkm4@cluster0.rdzqu.mongodb.net/devcamperDB?retryWrites=true&w=majority&appName=Cluster0
