--- devCamper-API/config/db.js
@@ -1,9 +1,0 @@
-const mongoose = require('mongoose');
-
-const connectDB = async () => {
-  const conn = await mongoose.connect(process.env.MONGO_URI);
-
-  console.log(`MongoDB connected.`);
-};
-
-module.exports = connectDB;
