--- devCamper-API/controllers/bootcamController.js
@@ -1,9 +1,0 @@
-// @desc    Get all bootcamps
-// @routes  /api/v1/bootcamps
-// access   Public
-
-const getAllBootcamps = (req, res, next) => {
-  res.json({ success: true, msg: 'Send the all data', hello: req.hello });
-};
-
-module.exports = getAllBootcamps;
