+++ package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
+    "mongodb": "^6.8.1",
     "mongoose": "^8.6.1",
     "morgan": "^1.10.0"
   }
